refactor(harvest): TriggerManager is now in charge of saving accounts 🛠
As AccountCreator will have to keep created account to be able to deal
with error, it will be almost identical to AccountEditor. The only
difference between them will be the way they save account. So we can
merge this logic into Trigger Manager and get rid of both AccountEditor
and AccountCreator

About TriggerSuccessMessage: this kind of visual confirmation should be
done in a parent component, for now in Cozy-Home. It will be easy to
get it back when we will need it again here.