feat(app-publish/downcloud): Use space name in remote path
Use the space name in the remote upload path to avoid overwriting files
when publishing an app with the same version and same slug to two
different spaces.

Currently, if I publish "app" v1.2.3 without specifying a space name,
it's uploaded to `www-upload/app/1.2.3/app.tar.gz`. Then if I publish
"app" v1.2.3 and I specify the space name "space", it's uploaded to the
same directory on downcloud. So it overwrites "app" v1.2.3 archive for
the default registry space.

With this commit, if a space name is provided, it's used to determine
the directory in which the archive should be uploaded. So I will upload
to `www-upload/app/1.2.3/app.tar.gz` and
`www-upload/space/app/1.2.3/app.tar.gz`.