feat(harvest): Add KonnectorJob ✨
KonnectorJob is a model which encapsulate existing
KonnectorJobWatcher and KonnectorAccountWatcher,
aiming to merge them into it in a near future.

For now we prefer to not introduce regressions
into TriggerManager, event if it will also
need a refactoring to use directly KonnectorJob.

KonnectorJob manages trigger, account and job
logic already existing into TriggerManager,
which is not the place they belong to.