fix(TriggerManager): Handle existing account without cipher update
When an account exists but is not linked to a cipher and we update it,
we can't find any cipher for it in the vault. In that case, we assumed
that we would find a cipher and did some things on it that crashed when
there was actually no cipher.