fix(cozy-auth): Don't redirect everytime we loged in
BREAKING CHANGE: Each time we try to reconnect from localStorage, we redirect
to loginPath (defaulted to /). It works great for application
that doesn't have to handle deeplinking for instance, but not
for them.

Since we can't be sure that `handleDeeplink` is called after `handleAfterLogin`
we can have a race condition between the both method resulting to redirecting
the user to loginPath instead of the url provided by the deeplink.

To prevent this issue, we should not redirect to loginPath in cozy-auth and
let the application does what she wants after a loggin.

In order to upgrade your app your need to :
- remove loginPath props
- in the `afterAuthenticated` callback you need to manually push to your previous
loginPath (ie router.push('/afterLogin'))