feat(cozy-harvest-lib): Cipher relationship should not be an array
In https://frama.link/pMf_cs4L we introduced an array in the
relationship. We should actually have just added the `data` attribute,
but keep an objet in it. With an array, it wrongly indicates that an
account can have a relationship to multiple ciphers. This is not the
case, it can have a relationship to only one cipher.