fix(cozy-harvest-lib): Always show account form when editing an account
When the user tries to edit an account, we show the form directly. But
ciphers are being fetched in background, and we were redirecting to the
ciphers list step as soon as we detect that some cipher exists.
Actually, when the user is on the account form step, we shouldn't
automatically redirect. If the user wants to see the ciphers list, he
can by clicking the "back" arrow.