fix: Bad password matching would result in null fields
Since the sort does not specify a number, false or true are used and
falsy values come first, this mean the search can find a cipher with
a different password that what is searched. Another check is necessary
if no cipher with the same password have ever been added.

For a reason I still ignore, this resulted in ciphers with null values
after both ciphers with different password were merged because the
deciphering would not work (mac failed error)