fix(Harvest): Change prod/sandbox environment detection
We previously detected production environment with production/staging
domains (mycozy.cloud, cozy.rocks) and defaulted to sandbox.
The detection is now inverted, we detect dev,int domains and set them
to sandbox and default to production.
This is a temporary fix, waiting for an api allowing an app to
properly detect it's environment.