fix: Shared Element when the document has no path
We assumed that a document has a path or at least that
we can fetch it. It works well with an io.cozy.files but
not with an io.cozy.albums since an album doesn't have a
path.

Instead of displaying nothing if there is no path, let's just
not make the call to the methods that need the path to work.