feat: Show confirmation prompt before prepublishing
Having the prompt after prepublishing causes several problems

- it means that we have to remember to go and say yes after the app has
  been uploaded to downcloud
- if there is an error in the detected slug / version, the app already
  has been uploaded with an error on downcloud

Here the prompt is shown as early as possible, meaning the developer
can go about its business after confirmation. Since we do not know the
appBuildUrl before prepublishing if it has not been passed, the logging
of its value is kept after prepublishing, along with the shasum.