feat: Re-organize EditContract code for buttons to be in correct layout
For the Dialog to be correctly layouted (with the buttons at the bottom)
on mobile, the Dialog components must rendered at the root of the
Dialog. Here, the structure is simplified (FormControls is removed) and
the DialogContent is now part of the EditContract page. Since on banks,
we'll use the same Dialog and not a standalone page, we do not need the
FormControlsWrapper and TitleWrapper.