feat: Ability to pass the triggers prop directly to KonnectorAccounts
Otherwise the KonnectorAccounts relies on the fact that triggers are
inserted into their respective konnectors by the custom store of the
home. This cannot be done easily when using cozy-client.

This is why here, we let KonnectorAccounts get the related triggers
directly with a prop. With this method, the user of KonnectorAccounts
must take care not to pass triggers that are unrelated to the konnector
passed in props, otherwise undefined behaviors will occur.