feat: Add softDeleteOrRestoreAccounts service
This service is useful to deal with soft delete and restore from
bitwarden, used for a trash feature.

When a cipher is soft deleted, a `deletedDate` field is added to the
cipher. Then, the service removes the encrypted credentials from the
referenced account and delete the associated trigger.

When the cipher is restored, the `deletedDate` is removed: the service
recreates the account's trigger. The account's credentials are then
restored through the `updateAccountsFromCipher` service.