fix: Handle cases where io.cozy.accounts does not have contract
If an account does not have any contract (for example because its
konnector does not fill its contracts relationship), we should not
fail completely. Here the code has been refactored so that the
SyncAccountSwitch also renders its label, and so that it renders null
if the related contract cannot be found in the account.