feat: Add redirectSlug to OAuthWindow
Since https://github.com/cozy/cozy-stack/pull/2674, the stack accepts
a new "slug" parameter. This param is used to redirect the user on the
specified slug after the end of the OAuth flow.

Before, the user was always redirected to the Home app, but now that
several apps (internal & external) use Harvest directly, we want to
redirect the user on the right app to provide a good UX.