feat(Harvest): Reuse unlinked accounts
Some accounts are not linked to any trigger anymore.
This is due to old bugs or when the user closes his window
too soon.

This PR will reuse these accounts, related to the same connector (via
account_type), fill them with new identifiers and link them to a new
trigger.