fix: Banking connector deletion
When the connector policy does not need to save the cipher in the
vault, we do not need to display the vault unlocker when deleting
the concerned cipher from the vault. This is the case for example
for banking connectors.

The problem here was that without the fix, the VaultUnlocker would
be displayed, but it returned null because of the saveInVault
condition in the VaultUnlocker render code, thus we could not delete
the account from a banking connector