feat: Main harvest modal is hidden when vault unlock is opened
Previously, when the vault unlocker was displayed, it was displayed over
the main harvest modal. The correct hiding of the main modal behind
the vault unlocker was difficult, and we could not use a narrower
VaultUnlocker since it would not hide completley the main modal. The
problem was then to hide the main modal when the vault unlocker is
displayed.

The structure of the code made it difficult as the vault unlocker was a
child of the main Harvest modal. Hiding the main modal would hide the
vault unlocker :/

Here, the code is reorganized so that TriggerManager is not responsible
to display the VaultUnlocker. Instead, it asks via its context, a
component near the root of Harvest (VaultUnlockPlaceholder) to display
the VaultUnlocker when necessary. At the same time, the HarvestDialog
checks, via the same context, if the VaultUnlocker is shown and hides
itself if true.

BREAKING CHANGE: When using the TriggerManager outside the main Harvest
Routes, you need to
`import { IntentTriggerManager } from 'cozy-harvest-lib'`, instead of
`import { TriggerManager } from 'cozy-harvest-lib'.