fix: Deletion of account
Due to the structure of the tree, we need to first close the modal then
delete the account. Otherwise, the account is deleted from the local
client state, resulting in account={null} being passed down the tree,
causing propTypes errors. Here the modal is closed optimistically, and
an alert is shown to the user to acknowledge the deletion.