fix: Only check if vault is locked if konnectorPolicy says to save in vault
We had an exception when the TriggerManager was not wrapped in a vault
context. It did not matter when the konnector that were dealt with did
not need to be saved in the vault (banking connectors for example).

After the fix, we do not need to wrap the TriggerManager in a vault ctx
if we are only dealing with konnectors that do not need to save in the
vault.

At the same time, crash with a helpful error if we deal with konnectors
that need to be saved in the vault while the TriggerManager has not
been wrapped in a VaultContext