fix: Update slide height correctly
Context:

    the slides of the harvest konnector modal are in a swipeable
    views.
    content of slides can be dynamically computed
    for slides of swipeable views to take the height of the current
    child, we need to set animateHeight to true and watch for
    subtree dom modifications

Bug:

A big white block was visible in the data tab when the content
of the configuration tab was much bigger (4). This is because a previous
commit removed the dom watching (necessary for 2) and thus the height
was not set dynamically (see 3).

The previous commit had removed updateHeight and animateHeight to
solve a bug where the content of the configuration was cropped
only when there was an error for the konnector. The fix meant
that the swipeable wrapper took the height of the tallest
slide, solving the cropping but introducing the bug 4.

To properly solve the bug, it is necessary to reintroduce the
animateHeight, the subtree modifications watching, and figure out
why swipeable views would compute the wrong height leading to cropped
content.

The cropped content appeared only when an error was shown.
The root cause was that the TriggerErrorInfo used
a margin-top to not be stuck to the tabs above. This margin was not
taken into account into the offsetHeight of the parent element
(only padding is taken into account) : swipeable views did not take
this margin when computing the height of the slide, leading to
cropped content.