feat(cozy-doctypes): Avoid a log for each transaction
In some cases (ex nestor connector with many transactions), the banking
reconciliator can generate a lot of logs, which is suspected to case
memory errors in connector's worker.

This logProgress added option allows to not log each transaction with
the `null` value and this option can be chosen by the connector.