fix(Harvest): Recreate non existing BI connection saved in account
In case of outdated, non existing connection id defined in the account.
The connection would fail with an unknown error caused by harvest trying
to update a non existing bi connection -> 404

This case is not supposed to happen but still happend because we do not
handle a double decoupled 2FA, which will be fixed in another PR. And
the user is blocked trying to authenticate his bank account and failing
miserably.

Now, we first check if the connection id is valid and create a new
connection if the connection id is not valid