feat: Add flag to activate files datacard
The file datacard functionality is behind the harvest.datacards.files
flag. This flag activates the file datacard for konnectors that have
permission on io.cozy.files and deactivates the drive card that is
shown for triggers that have message.folder_to_save.

The precise logic on when to show the datacards has not been completely
decided yet. It will surely change in the future.