fix: When no number or null, use id attribute
In the Natixis example, there is no value
in the number attribute of the account.

And so when matching the accounts the score
was 0 (no matching).
As a result we had duplicate accounts.

A “rule” has been added which allows to take into
account the case where the "number" attribute
is null.

In this case we use the account ID which
allows for a correspondence.