fix(Harvest): SelectBox use position Fixed
In previous attempt to handle the overflow, we
decided to calculate the size of the select box
by measuring the size of the Select's container
and calculating the available space.

See https://github.com/cozy/cozy-libs/pull/240

But if the SelectBox is near to the bottom of the
container, its size is not big enough.

To fix that issue, we use the fixed position of
ReactSelect to overflow the content and use a
default max-height to 300px.

Since the "container" props was added by the PR
we "revert", I removed them.