fix: esc touch does the same as the back button.
The bug was: after opened the viewer, if I touch the "ESC" we close the
"first modal" aka the Harvest one (containing herself this Overlay).

We can pass an `onEscape` props on the `Overlay` component, but in fact
Overlay is below the `Harvest Modal` as you can see we cheated with a
zIndex to display it in top. Since the Modal is "first", then when we
espace, we close the modal (and since the modal contains the Overlay
it closes all the stuff).

To fix that, we use the Modal component from M-UI. Like that, M-UI
handles the "stack component" for us, so the latest opened modal get the
priority for events.

I now import Modal from m-ui itself, but like that I do not need to
update UI. Will do that later.