fix: Label height for Account Form
Before, we set the fontSize to 0 in order to not display the
ObfuscateChar. But by doing so, I do know why, but the
height of the label is not the good one at least on Chromium's
based browser.

To avoid this issue, let's use display none. It will be easier
for Password manager to detect our trick, but we already have
an aria-label saying the same thing. So ...