fix: Harvest Konnector translation
Relying on of non "lifecycled" react variable
is pretty always a bad idea.

Here, we have a bad sync between runtime variable
and react state.

Now the I18n stuff is unmounted when we close the
Modal, but we didn't remove the konnector from
the already loaded konnector. The result was that
when we opened again the modal after having
closing it once we didn't have the right translation.

Let's remove this stuff.