fix: Handle undefined this.konnector for OAuth connectors
In ConnectionFlow, this.konnector is undefined.
I don't see any justification for that but it won't have any
other nagative effect on the connector exection.
This causes a fatal error on all applications using
harvest and trying to configure an OAuth connector

This is a hotfix.
We will check why this variable is undefined in later investigation