fix: Change timeout to get BI temporary token
In some (too many) case, BI can take more than 30s to give response.
This leads to timeout error message to the user which occur in their
first connection or when renewing their connection.
And since these errors occur in the browser, they are not visible
in our usual connector error dashboard.

We saw that when BI is long to response, it takes usually 35s to 40s but
in most cases it takes 2s to 3s