feat: Add a manual typescript declaration file
This will be build upon as needs arise.
For now, I only need this method to be typed.
This is better to do it here rather than in my other app in my opinion.

The tsconfig.json file is only added for ESlint purposes.
Without it ESlint will throw because he found a ts file but no tsconfig.
The config in itself currently has no other purpose.