fix: VaultClient can be undefined at first
Since the 4.1.0 of cozy-keys-lib, we
render children even if vaultClient is
not defined yet. In that case we we were
displaying TriggerManager without vaultClient.
It was raising an error.
The current fix, is to not display the
TriggerManager when vaultClient is null.

The result for the end user, was to get a blank
form when configuring a konnector from the store