feat: Add the ability to display multiple contacts per file
Now that we can have multiple contacts per file,
we need to display them correctly,
without visually duplicating the files involved.

The UI rules are as follows:
- A section containing the names of the contacts is created in the list.
- If there are more than 2 contacts,
then "..." is added after the contact names.
- If the names of the contacts are identical
(and there are only 2 of them),
then they are merged (e.g. `Isabelle and Paul Durand`).