feat: Add a synchronization button after Contract list
This button will show a BI webview which allow to modify contract
synchronisation.

The webview can be shown with a popup in web environment or with an
InAppBrowser in flagship application environment.

Restriction : at the moment, the BI webview does not allow to edit only
one BI connection, all the connections of the user with will be visible
and editable. This is not acceptable but will be fixed later.

Information : this new button is only displayed if the current connector
policy implements the fetchContractSynchronizationUrl method, and this
method is available only when the harvest.bi.webview flag is set to
true. Then no current screen of the home application will affected by
this change if the flag is not set.