fix(cozy-intent): Allow overriding existing connection
By being too strict when initiating a webview connection,
we made it impossible to handle some edge cases scenarios
who actually need to erase their old session and create a new one
in the same messenger object. Simplest way is to let them
write the new connection over the old one,
indeed the old connection will never be used again
This change was motivated by cozy-notes,
this application does some back and forth between drive and note,
the context was lost and it was hard to recreate a handshake.
If we let the app seek a new context by itself when it renders,
this issue is handled automatically.
Hopefully this doesn't create side effects