fix: Fix width flickering in ModalHeader
The header has 2 width.
One with children, one without children.
It was checking for undefined to decide if it had children or not.
Undefined is not a correct value for a React children,
therefore we now check for the null value.
This allows us to know in sync if we must display the big or the small
icon container.