feat: Handle user dismiss in Oauth window
The dismiss or cancel action of the user in the oauth window generates
an error message which was not handled by Harvest.

Now, a dismiss action of the user causes the closing of the popup or
InAppBrowser.

This supposes that https://github.com/cozy/cozy-stack/pull/3445 is
merged in cozy-stack to not cause a redirect to the store application.