fix: OauthWindowInAppBrowser re-render
The component was working good when we had an intentsApi provided but
not when we were using the webviewIntent.

This is because the hook was executed several times since at each render
we created a new method for fetchSessionCode since we were creating it
by doing const fetchSessionCode = () => webview.call().

I tried to mutualize the code, but I failed during the small amount of
time I had to fix this issue.