fix: Remove a reconnection step
When the trigger is in a state which needs reconnection and when the
user clicks on the "Reconnect" button, he still needed to click on a
"reconnect" button one more time to display the BI webview.

Now the BI webview is displayed directly.

There are still some screens we don't want to display after the BI
webview but this will be fixed in another PR