feat: Detect reconnect explicitely
Before, we detected a reconnect with the presence of a connection id in
the account. But if we use the creation BI webview to handle
synchronizations, there will be a connection id in the account and the
wrong url parameters will be used.

Now we pass a "reconnect" parameter explicitiely