feat: Use BI account creation webview to handle accounts synchonization
The goal is to avoid to display the BI manage webview which allows
modify, delete or create other BI connections of the user, which we
cannot handle on cozy side at the moment.

The user will need to enter it's own bank credentials before modifying
accounts synchronization.

The webview will close on its own.