fix: Dispatch account deletion after BI connection deletion
This store update will espacially allow banks to update the
connected/disconnected status of it's accounts without adding realtime
on accounts, which causes more troubles for OAuthWindow component which
will unmount.

This is a quick fix. A better approach is in progress to make
OAuthWindow independent of Harvest components tree