fix(mespapiers): Group papers by connector identifier
We initially grouped by `sourceAccount` which is the id of the
account used by a connector. However, if we disconnect an account
from a connector, it is deleted. So when you reconnect, you create
a new account, with a new `sourceAccount`. So we can't gather the
documents according to this. We now rely on the slug of the
connector and the `sourceAccountIdentifier` (which is usually the
login of the account) to avoid gathering documents with the same
login but from different connectors.