feat(harvest): Replace router dependency in component by context dep
We want harvest to work without major changes with both a v3
router (Banks and Drive), v4 (use case "Home"), and with a v6 router
(use case "MyPapers").

Initially we didn't want to have a backward compatibility with v3 and
Banks because we thought it was autonomous and passed
everything to harvest. But this is not the case. Moreover we initially
thought of using a context to determine which version of the router
to use, but this poses problems because there are several "entry
points" to Harvest and we are not sure to have the context in all
cases.

The idea is that we remove the direct dependency on the router via
`withRouter` in the components, and replace it with a global
`withAdaptiveRouter` dependency that will use either version of the
router.

Some components were already using this semblance of an
approach with the approach with `withMountPointHistory`. So we
improved it by adding `location` to this HOC.