fix(harvest): LaunchTriggerAlert text vertical alignement
A better way to do this would be to initially see in the mockups that
the Alert component was not used as proposed by cozy-ui. Since this
is an isolated case (and we don't have time to review the specs and fix
it on the cozy-ui side) for now let's do an application side fix.