feat: Update label-slugs.js file
This avoid a degraded banking accounts reconciliation when account
institution label is unknown.

We should should find a better way of reconciliation not dependent on
institutionLabels but this will probably avoid many unwanted accounts
merge at the moment.

The previous associations were also made to recognize old linxo accounts
but this is not in the scope anymore