feat: Add cozy-tsconfig lib
Currently copy pasting this config a lot everywhere.
Seems a lot better to just add it as a lib, that way we can control
TS globally in our apps.

Please note that this config does NOT transpile files to javascript.
It's only meant to be used as a declaration file generator,
even for JS files.