feat(realtime): Add _type on returned doc by the handlers
in the document returned by the realtime, there is no _type
associated with the doctype, which requires some gymnastics
on the app side when the realtime is connected to different
doctypes.

Thanks to this improvement, it is enough to analyze the document
to know to which doctype it is associated.

It is in the case where the same callback is connected on different
subscribes, we do not discriminate upstream but downstream
via the doc.

The other reason is also to standardize compared to the return
of CC which normalizes the doc and adds this _type.