fix: Reconnect WebSocket
If an application is opened for a long time, aka in a
long life browser's tab and you never close the browser
and that the client can go on a sleep or else, WS can
not be connected.

So we try 2 things here:
- First is to reconnect after a refreshedToken. It can
be useful because CozyClient will try to refresh the token
after 24h. So if the token is refreshed, it means that
the client has an internet connexion and just make a request
so it's the right time to reconnect with this new token

- Second is to play with the visibility API. Before, the
reconnexion was not very aggressive. We only tried if we
were in a process to trying to reconnect by accelerating
the reconnexion process. Now we try to reconnect also if
the current websocket is closed.