fix: Display harvest banner when have multiple konnectors
with same qualification_labels.

There can be several konnectors
with the same qualification_labels (ex: pay_sheet).
In this case, the harvest banner must appear on
each file list of each account

`paperslistByContact` is the array of aggregated data
for each contact or account,
it seems normal to add the associated konnector info