feat: Return also readOnly status from getSharedDocument
The `getSharedDocument` function is only used in cozy-drive at the moment.  Cozy-notes has its own version that comes close to these changes. It may be possible to use the same for both applications in the future.

BREAKING CHANGE: The function now returns the id encapsulated in an object. You can extract the id like this `const { id } = getSharedDocument(client)`