fix: Refetch current trigger when it is running
When creating a new account/trigger couple, when the given trigger has
started running, the FlowProvider is reloaded, which means it
unsubscribes from realtime events until it is reloaded.

If the konnector execution is really fast (example: dummy-> 1s), its
execution could end during the FlowProvider reload (which is a little
less than 1s on my development environment).

To fix this issue, I force the ConnectionFlow to refetch the current
trigger if this trigger is in the state "running", in case we missed the
konnector execution end.