fix(harvest): DialogBackButtonon config modal
This will fix an issue where the browser could manage to go back from
that modal, but the UI component backButton could not,
This bug was caused by the fact that the onClick callback
was using both the wrong method (push instead of replace) and the
wrong redirect path.

Though do we want a replace or push here? I must say
both options have merits