fix(mespapiers): Link to install konnector
The fact of going through an `onClick` on a `li` tag,
the opening of the link was obligatorily done in a new tab.
The fix is to go through an `a` tag
and pass it the target attribute if the flag is present.
Thus we find the initial and expected behavior