fix: Avoid FlowProvider reload on trigger launch
With harvest.inappconnectors.enabled flag, when launching a trigger,
there is a flicker which is caused by the FlowProvider component reload.

To fix it, we change the component tree structure to avoid the rerender
of the AccountModal component.

This fix has been tested in home application with and without the flag
enabled and also in the Mes Papier application.