fix: Display reconnection webview for powens konnectors
Even with harvest.inappconnectors.enabled flag activated.

The difference with previous interface is that the "reconnect" button
and the synchronize button have merge.

When the konnector is a powens one and when the error is solvable via
reconnection, the reconnection action (display the powens webview) is
done instead of just launching the konnector.

The reconnection button has exactly the same style as the
synchronization button.