fix: Realtime issue on Harvest
We had issue with the realtime when closing some Harvest
component.

We send an unsubscribe message to the server even if we
still have subscription.

This was done because we didn't share the realtime instance
between the app and harvest for this component.

So when the component was unmounted, cozy-realtime was
checking if there was any subscription remaining to this
doctype and then send the message to close the channel.

Since we had several instance of cozy-realtime, its enough to
unsubscribe if one of the instance send the unsubscribe method.

In order to avoid this kind of side effect, we should always
use the singleton of cozy-realtime provider by cozy-client.

There is still a few new CozyRealtime in Harvest, we should take
care of them also.