feat: Expose native intent service
On flagship app, we need to call a intent from a webview inside an
intent in the flagship app. To do so, we need to access the native
intent service. We can not use useNativeIntent webhook, because in
the flagship app we are not in a React component, and we are at the
same level of the provider definition.