feat(minilog): Recognize ReactNative as modern env
By modifying the library code in this way,
we're ensuring that the more capable
logging backend console.minilog is used
not only when the code is running in a Chrome browser,
but also when it's running in a React Native environment.
This ensures better compatibility with different running environments
without the need for users to modify their global objects
(like navigator), resulting in cleaner and more maintainable code.