fix(mespapiers): File replacement confirmation
When creating a paper, if the file comes from the Cozy
when you click submit, a modal must appear to ask
if you want to replace the selected file or not.

In commit 8e14d14, having normalized the File type
and no longer using the Blob type, introduced a regression.
These files are no longer recognized
and the modal no longer appears. (which keeps the 2 files)