feat(mespapiers): Add `SubmitButtonComponent` to `ContactDialog` step
The initial need being to have a `note` step last,
after the `contact` step,
we need to make the action of the `contact` step conditional.
This step must also be able to go to the next step if it exists.

At the moment we don't want to add
this possibility/complexity to the other steps.