feat(mespapiers): Remove event listener on note content copy
It was not working on desktop.

After a lot of research and try, I didn't
find a better way to have something working on mobile and desktop.
I don't know why but fetchJSON approach is only working on desktop
and useFetchJSON only on mobile. I don't think the cozy code is to blame, because with useFetchJSON or fetchJSON we get the content of the note as expected, but it's the copy-text-to-clipboard lib that doesn't seem to be working despite receiving the content.