feat(mespapiers): Remove the `stepIndex` property from the config file
This property was added when the project was created.
Its main purpose was to guarantee the order of the steps.
Which is useless in JS because the order of JSON arrays
is guaranteed by the language.

This will make it easier to manipulate the steps in
addition to making them lighter.