feat(cozy-intent): Avoid localMethods rerender
Previously, the context was recreated each time the localMethods
reference changed.
Now, if this object change, we will just update the existing
context instance with this new localMethods object.
This will optimize the behaviour of the provider because it avoids
destructive rerenders that are not necessary for the intents
to still work and use the updated localMethods with the existing service