fix: Index only existing document after a realtime update
When a new share is made, realtime trigger the indexation of all documents in the new sharing as well as the shortcut. If the recipient has not synchronised the sharing, the document does not exist, leading to errors in the indexing script. This is why we only index files if the share is ready, so that the files exist