fix: BuildSharingLink works also for perm from mes papiers
We used to use `email` as `codes` attribute for a sharingByLink.
But when we use cozy-client to create a Permission, by default
the codes attribute is set to `code`. MesPapiers app is using this
default behavior... So the sharing by link created by mes papiers
didn't appear correctly in cozy-sharing.
This is a bit ugly, we should have a better way to know if this is
a sharing by link or not.